Move environment fixture to root
